refactor(session-recovery): split recovery hook facade
